# swagger.json
{"openapi":"3.0.0","paths":{"/mass":{"post":{"operationId":"MassController_create","summary":"Create mass id","description":"Explain more about mass id create endpoint","parameters":[{"name":"x-api-version: 2","in":"header","description":"Api version","schema":{"type":"string"}},{"name":"x-api-token","in":"header","description":"Api access token","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMassDto"}}}},"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMassResponse"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMassError400Response"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMassError401Response"}}}}},"tags":["Mass"]},"get":{"operationId":"MassController_findAll","summary":"List masses ids","description":"Explain more about mass id create endpoint v1","parameters":[{"name":"x-api-version: 2","in":"header","description":"Api version","schema":{"type":"string"}},{"name":"x-api-token","in":"header","description":"Api access token","schema":{"type":"string"}}],"responses":{"200":{"description":"Success","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CreateMassDto"}}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMassError400Response"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMassError401Response"}}}}},"tags":["Mass"]}}},"info":{"title":"ZWF API","description":"External Mass Api","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateMassDto":{"type":"object","properties":{"currentValue":{"type":"number","description":"A number representing the current value of the Mass. Must be a positive number","example":100},"externalCreatedAt":{"format":"date-time","type":"string","description":"A date representing the external creation date of the Mass","example":"2023-04-12T12:34:56.789Z"},"externalId":{"type":"string","description":"An optional string representing the external ID of the Mass","example":"ext123"},"id":{"type":"string","description":"A string representing the unique ID of the Mass. Must not be empty","example":"123456"},"initialValue":{"type":"number","description":"A number representing the initial value of the Mass. Must be a positive number.","example":50},"isPublic":{"type":"boolean","description":"A boolean indicating if the Mass is public or not.","example":true},"isPubliclySearchable":{"type":"boolean","description":"A boolean indicating if the Mass is publicly searchable or not.","example":true},"parentMassId":{"type":"string","description":"An optional string representing the ID of the parent Mass object.","example":"123456"},"type":{"type":"string","description":"A string representing the type of the Mass. Must not be empty.","example":"parent123"},"subtype":{"type":"string","description":"A string representing the subtype of the Mass. Must not be empty.","example":"type2"}},"required":["currentValue","externalCreatedAt","externalId","id","initialValue","isPublic","isPubliclySearchable","type","subtype"]},"CreateMassResponse":{"type":"object","properties":{"id":{"type":"string","description":"A string representing the unique ID of the Mass","example":"123456"}},"required":["id"]},"CreateMassError400Response":{"type":"object","properties":{"statusCode":{"type":"number","example":401},"message":{"type":"string","example":"Invalid Payload"}},"required":["statusCode","message"]},"CreateMassError401Response":{"type":"object","properties":{"statusCode":{"type":"number","example":401},"message":{"type":"string","example":"Unauthorized"}}}}}}